requirements.txt
# Core ML/Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM & Agents
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=1.0.0
langgraph>=0.2.0

# RAG
chromadb>=0.5.0
sentence-transformers>=2.2.0

# PDF Processing
pymupdf>=1.23.0

# Experiment Tracking
mlflow>=2.10.0

# UI
streamlit>=1.30.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0
